requirements.txt
fastapi
uvicorn
python-multipart

pymupdf
python-dotenv
tiktoken

requests

google-genai
ddtrace>=4.3.1
openai>=1.0.0
